Sync API Output: ConfirmationResponse
