+++ src/main/java/teammates/ui/webapi/action/ConfirmFeedbackSessionSubmissionAction.java
@@ -10,7 +10,8 @@ import teammates.common.util.Const;
 import teammates.common.util.EmailSendingStatus;
 import teammates.common.util.EmailWrapper;
 import teammates.logic.api.EmailGenerator;
-import teammates.ui.webapi.output.ApiOutput;
+import teammates.ui.webapi.output.ConfirmationResponse;
+import teammates.ui.webapi.output.ConfirmationResult;
 
 /**
  * Confirm the submission of a feedback session.
@@ -102,34 +103,5 @@ public class ConfirmFeedbackSessionSubmissionAction extends BasicFeedbackSubmiss
         }
 
         return new JsonResult(new ConfirmationResponse(ConfirmationResult.SUCCESS, "Submission confirmed"));
-    }
-
-    /**
-     * The result of the confirmation.
-     */
-    enum ConfirmationResult {
-        SUCCESS,
-        SUCCESS_BUT_EMAIL_FAIL_TO_SEND
-    }
-
-    /**
-     * The output format of {@link ConfirmFeedbackSessionSubmissionAction}.
-     */
-    public static class ConfirmationResponse extends ApiOutput {
-        private final ConfirmationResult result;
-        private final String message;
-
-        public ConfirmationResponse(ConfirmationResult result, String message) {
-            this.result = result;
-            this.message = message;
-        }
-
-        public ConfirmationResult getResult() {
-            return result;
-        }
-
-        public String getMessage() {
-            return message;
-        }
     }
 }
